feat: Create custom hook for post list components and refactor post list components
